KM-17647: Move consent-preservation test out of AccountTests, which CI skips entirely
